pyproject.toml
[project]
name = "mero-lagani"
version = "0.1.0"
description = "Service ready to provide Latest IPO news, and Automation bot which fills up the IPO quota"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=6.0.2",
    "djangorestframework>=3.16.1",
    "dotenv>=0.9.9",
    "selenium>=4.40.0",
    "webdriver-manager>=4.0.2",
    "django-redis>=5.4.0",
    "redis>=5.2.1",
    "python-dotenv>=1.2.1",
]
